requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
streamlit
python-dotenv
requests
beautifulsoup4
rank-bm25
pypdf
trafilatura
sentence-transformers
tiktoken
torch
pandas